remove old modal
