+++ client/packages/common/src/hooks/useFeatureFlags/useFeatureFlags.ts
@@ -34,8 +34,5 @@ export const useFeatureFlags = () => {
   return {
     // Enable the new demo table, with better usability
     tableUsabilityImprovements: !!featureFlags['table_usability_improvements'],
-    // Enable the new demo create stocktake modal (see https://github.com/msupply-foundation/open-msupply/issues/8773)
-    createStocktakeModalUsabilityImprovements:
-      !!featureFlags['create_stocktake_modal_usability_improvements'],
   };
 };
