+++ client/packages/inventory/src/Stocktake/ListView/AppBarButtons.tsx
@@ -10,14 +10,12 @@ import {
   useToggle,
   useSimplifiedTabletUI,
   useExportCSV,
-  useFeatureFlags,
 } from '@openmsupply-client/common';
 import { useTranslation } from '@common/intl';
 import { useStocktakeOld } from '../api';
 import { stocktakesToCsv } from '../../utils';
-import { CreateStocktakeModal } from './CreateStocktakeModal';
 import { CreateStocktakeInput } from '../api/hooks/useStocktake';
-import { NewCreateStocktakeModal } from './NewCreateStocktakeModal';
+import { CreateStocktakeModal } from './CreateStocktakeModal';
 
 interface AppBarButtonsProps {
   description: string;
@@ -52,8 +50,6 @@ export const AppBarButtons = ({
     exportCSV(csv, t('filename.stocktakes'));
   };
 
-  const featureFlags = useFeatureFlags();
-
   return (
     <AppBarButtonsPortal>
       <Grid container gap={1}>
@@ -62,23 +58,13 @@ export const AppBarButtons = ({
           label={t('label.new-stocktake')}
           onClick={modalController.toggleOn}
         />
-        {featureFlags?.createStocktakeModalUsabilityImprovements ? (
-          <NewCreateStocktakeModal
-            open={modalController.isOn}
-            onClose={modalController.toggleOff}
-            onCreate={onCreate}
-            isCreating={isCreating}
-            description={description}
-          />
-        ) : (
-          <CreateStocktakeModal
-            open={modalController.isOn}
-            onClose={modalController.toggleOff}
-            onCreate={onCreate}
-            isCreating={isCreating}
-            description={description}
-          />
-        )}
+        <CreateStocktakeModal
+          open={modalController.isOn}
+          onClose={modalController.toggleOff}
+          onCreate={onCreate}
+          isCreating={isCreating}
+          description={description}
+        />
         {!simplifiedTabletView && (
           <LoadingButton
             startIcon={<DownloadIcon />}
