+++ client/packages/inventory/src/Stocktake/ListView/CreateStocktakeModal.tsx
@@ -1,12 +1,14 @@
 import React, { useState } from 'react';
 import {
+  Alert,
   BasicSpinner,
-  Checkbox,
   DateTimePickerInput,
   DialogButton,
   InputWithLabelRow,
+  RadioGroup,
 } from '@common/components';
 import { DateUtils, useFormatDateTime, useTranslation } from '@common/intl';
+import { FormControlLabel, Radio } from '@mui/material';
 import { useDialog } from '@common/hooks';
 import {
   useStockListCount,
@@ -16,6 +18,7 @@ import {
   MasterListRowFragment,
   useMasterListLineCount,
   VVMStatusSearchInput,
+  VvmStatusFragment,
 } from '@openmsupply-client/system';
 import {
   Box,
@@ -25,7 +28,6 @@ import {
   usePreferences,
 } from '@openmsupply-client/common';
 import { CreateStocktakeInput } from '../api/hooks/useStocktake';
-import { VvmStatusFragment } from 'packages/system/src/Stock/api';
 
 const LABEL_FLEX = '0 0 150px';
 interface NewStocktakeModalProps {
@@ -37,12 +39,17 @@ interface NewStocktakeModalProps {
 }
 
 interface ModalState {
-  location: LocationRowFragment | null;
   masterList: MasterListRowFragment | null;
-  expiryDate: Date | null;
-  createBlankStocktake: boolean;
-  includeAllMasterListItems: boolean;
   vvmStatus: VvmStatusFragment | null;
+  location: LocationRowFragment | null;
+  expiryDate: Date | null;
+  itemStatus: ItemStatus;
+}
+
+enum ItemStatus {
+  All = 'all',
+  InStock = 'in stock',
+  None = 'none',
 }
 
 export const CreateStocktakeModal = ({
@@ -64,22 +71,14 @@ export const CreateStocktakeModal = ({
   });
 
   const [
-    {
-      location,
-      masterList,
-      vvmStatus,
-      expiryDate,
-      createBlankStocktake,
-      includeAllMasterListItems,
-    },
+    { location, masterList, vvmStatus, expiryDate, itemStatus },
     setState,
   ] = useState<ModalState>({
     location: null,
     vvmStatus: null,
     masterList: null,
     expiryDate: null,
-    createBlankStocktake: false,
-    includeAllMasterListItems: false,
+    itemStatus: ItemStatus.InStock,
   });
 
   const stockFilter: StockLineFilterInput = {
@@ -95,6 +94,7 @@ export const CreateStocktakeModal = ({
     vvmStatusId: vvmStatus && {
       equalTo: vvmStatus.id,
     },
+    hasPacksInStore: itemStatus === ItemStatus.InStock ? true : undefined,
   };
 
   const { data } = useStockListCount(stockFilter);
@@ -103,7 +103,7 @@ export const CreateStocktakeModal = ({
   const { localisedDate } = useFormatDateTime();
 
   const generateComment = () => {
-    if (createBlankStocktake) return '';
+    if (itemStatus === ItemStatus.None) return '';
 
     const filterComments: string[] = [];
 
@@ -158,13 +158,15 @@ export const CreateStocktakeModal = ({
       masterListId: masterList?.id,
       locationId: location?.id,
       vvmStatusId: vvmStatus?.id,
-      createBlankStocktake,
+      createBlankStocktake: itemStatus === ItemStatus.None,
       expiresBefore: Formatter.naiveDate(adjustedExpiryDate),
       isInitialStocktake: false,
-      includeAllMasterListItems,
+      includeAllMasterListItems:
+        itemStatus === ItemStatus.All && !!masterList?.id,
       description,
       comment: generateComment(),
     };
+
     onCreate(args).then(id => {
       if (id) {
         navigate(id);
@@ -172,16 +174,12 @@ export const CreateStocktakeModal = ({
     });
   };
 
-  let estimatedLineCount = 0;
-  if (createBlankStocktake) {
-    estimatedLineCount = 0;
-  } else {
+  const estimateLineCount = (): number => {
     const stockCount = data?.totalCount ?? 0;
-    estimatedLineCount =
-      includeAllMasterListItems && masterListLineCount
-        ? Math.max(masterListLineCount, stockCount)
-        : stockCount;
-  }
+    return itemStatus === ItemStatus.All && masterListLineCount
+      ? Math.max(masterListLineCount, stockCount)
+      : stockCount;
+  };
 
   return (
     <>
@@ -207,30 +205,14 @@ export const CreateStocktakeModal = ({
               <InputWithLabelRow
                 labelProps={{ sx: { flex: `${LABEL_FLEX}` } }}
                 Input={
-                  <Checkbox
-                    style={{ paddingLeft: 0 }}
-                    checked={!!createBlankStocktake}
-                    onChange={e =>
-                      setState(() => ({
-                        createBlankStocktake: e.target.checked,
-                        masterList: null,
-                        includeAllMasterListItems: false,
-                        location: null,
-                        expiryDate: null,
-                        vvmStatus: null,
-                      }))
-                    }
-                  />
-                }
-                label={t('stocktake.create-blank')}
-              />
-              <InputWithLabelRow
-                labelProps={{ sx: { flex: `${LABEL_FLEX}` } }}
-                Input={
                   <MasterListSearchInput
-                    disabled={!!createBlankStocktake}
+                    disabled={itemStatus == ItemStatus.None}
                     onChange={masterList =>
-                      setState(prev => ({ ...prev, masterList }))
+                      setState(prev => ({
+                        ...prev,
+                        masterList,
+                        itemStatus: ItemStatus.InStock,
+                      }))
                     }
                     selectedMasterList={masterList}
                     width={380}
@@ -238,37 +220,14 @@ export const CreateStocktakeModal = ({
                 }
                 label={t('label.master-list')}
               />
-              {masterList ? (
-                <InputWithLabelRow
-                  labelProps={{ sx: { flex: `0 0 250px` } }}
-                  sx={{ paddingLeft: '160px' }}
-                  Input={
-                    <Checkbox
-                      style={{ paddingLeft: 0 }}
-                      disabled={!masterList || createBlankStocktake}
-                      checked={!!includeAllMasterListItems}
-                      onChange={e =>
-                        setState(prev => ({
-                          ...prev,
-                          includeAllMasterListItems: e.target.checked,
-                          vvmStatus: null,
-                          location: null,
-                          expiryDate: null,
-                        }))
-                      }
-                    />
-                  }
-                  label={t('stocktake.all-master-list-items')}
-                  labelRight={true}
-                />
-              ) : null}
+
               <InputWithLabelRow
-                labelProps={{ sx: { flex: `${LABEL_FLEX}` } }}
+                labelProps={{
+                  sx: { flex: `${LABEL_FLEX}` },
+                }}
                 Input={
                   <LocationSearchInput
-                    disabled={
-                      !!createBlankStocktake || includeAllMasterListItems
-                    }
+                    disabled={itemStatus != ItemStatus.InStock}
                     onChange={location =>
                       setState(prev => ({ ...prev, location }))
                     }
@@ -282,10 +241,8 @@ export const CreateStocktakeModal = ({
                 labelProps={{ sx: { flex: `${LABEL_FLEX}` } }}
                 Input={
                   <DateTimePickerInput
-                    width="100%"
-                    disabled={
-                      !!createBlankStocktake || includeAllMasterListItems
-                    }
+                    width={380}
+                    disabled={itemStatus != ItemStatus.InStock}
                     value={expiryDate}
                     onChange={expiryDate =>
                       setState(prev => ({ ...prev, expiryDate }))
@@ -294,16 +251,13 @@ export const CreateStocktakeModal = ({
                 }
                 label={t('label.items-expiring-before')}
               />
-
               {manageVvmStatusForStock && (
                 <InputWithLabelRow
                   label={t('label.vvm-status')}
                   labelProps={{ sx: { flex: `${LABEL_FLEX}` } }}
                   Input={
                     <VVMStatusSearchInput
-                      disabled={
-                        !!createBlankStocktake || includeAllMasterListItems
-                      }
+                      disabled={itemStatus != ItemStatus.InStock}
                       onChange={vvmStatus =>
                         setState(prev => ({
                           ...prev,
@@ -319,9 +273,61 @@ export const CreateStocktakeModal = ({
 
               <InputWithLabelRow
                 labelProps={{ sx: { flex: `${LABEL_FLEX}` } }}
-                Input={estimatedLineCount}
-                label={t('label.stocktake-estimated-lines')}
+                Input={
+                  <RadioGroup
+                    value={itemStatus}
+                    sx={{ margin: '0 auto' }}
+                    onChange={event => {
+                      setState(prev => ({
+                        ...prev,
+                        itemStatus: event.target.value as ItemStatus,
+                      }));
+                    }}
+                  >
+                    <FormControlLabel
+                      value={ItemStatus.InStock}
+                      control={<Radio />}
+                      label={t('report.in-stock')}
+                    />
+                    <FormControlLabel
+                      disabled={
+                        !masterList || !!expiryDate || location || vvmStatus
+                          ? true
+                          : false
+                      }
+                      value={ItemStatus.All}
+                      control={<Radio />}
+                      label={t('label.all')}
+                    />
+                    <FormControlLabel
+                      disabled={
+                        masterList || expiryDate || location || vvmStatus
+                          ? true
+                          : false
+                      }
+                      value={ItemStatus.None}
+                      control={<Radio />}
+                      label={
+                        t('label.none') +
+                        ' (' +
+                        t('label.blank-stocktake') +
+                        ')'
+                      }
+                    />
+                  </RadioGroup>
+                }
+                label={t('label.item-status')}
               />
+
+              {itemStatus == ItemStatus.None ? (
+                <Alert severity="success">
+                  {t('message.create-blank-stocktake')}
+                </Alert>
+              ) : (
+                <Alert severity="info">
+                  {t('message.lines-estimated', { count: estimateLineCount() })}
+                </Alert>
+              )}
             </Box>
           ) : (
             <Box sx={{ height: '100%' }}>
