+++ client/packages/system/src/Stock/index.ts
@@ -5,6 +5,7 @@ export {
   useRepack,
   useVvmStatusesEnabled,
   StockListParams,
+  VvmStatusFragment,
 } from './api';
 export * from './ListView';
 export * from './DetailView';
